requirements.txt
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.2.10
asn1crypto==1.5.1
autopep8==2.3.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-crum==0.7.7
django-heroku==0.3.1
django-ipware==3.0.2
django-widget-tweaks==1.4.8
filelock==3.0.12
flake8==2.3.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.10
legacy-cgi==2.6.2
lxml==5.3.0
mccabe==0.6.1
oscrypto==1.3.0
pep8==1.7.1
pillow==11.0.0
psycopg2==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==2.2.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
PyPDF2==1.26.0
python-bidi==0.6.3
python-decouple==3.4
pytz==2021.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.4.1
svglib==1.5.1
tinycss2==1.4.0
toml==0.10.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.16
